updated csBehavior_v9
Figured out why the gui wasn't working with the opto task. Fixed some bugs related to opto timing parameters. added some opto options on the main gui window for my convenience, for a specific set of experiments (but this can always be removed/changed). Still need to add in the channel change for masking stim. I need to test this on the 2p computer tomorrow because both the discrimination and opto task were freezing on my laptop. And I'll email you the flyback version in a sec!